% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{binom_ci}
\alias{binom_ci}
\title{Confidence Intervals for Binomial Probabilities}
\usage{
binom_ci(x, n, alpha = 0.05, method = c("wilson", "exact", "asymptotic",
  "all"), include.x = FALSE, include.n = FALSE, return.df = FALSE)
}
\arguments{
\item{x}{vector containing the number of "successes" for binomial variates}

\item{n}{vector containing the numbers of corresponding observations}

\item{alpha}{cprobability of a type I error, so confidence coefficient =
1-alpha}
}
\description{
An implementation of the \code{\link[Hmisc]{binconf}} function in Frank
Harrell's Hmisc package. Produces 1-alpha confidence intervals for binomial
probabilities.
}
\seealso{
\code{\link[Hmisc]{binconf}}
}
\keyword{internal}
