% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dissolved_Oxygen}
\alias{Dissolved_Oxygen}
\title{Dissolved oxygen measurements from the Tres Palacios rivers}
\format{A data frame with 236 rows and 6 variables:
\describe{
  \item{Station_ID}{unique water quality monitoring station identifier}
  \item{Date}{sampling date in yyyy-mm-dd format}
  \item{Param_Code}{unique parameter code}
  \item{Param_Desc}{paramter description with units}
  \item{Average_DO}{mean of dissolved oxygen measurement, in mg/L}
  \item{Min_DO}{minimum of dissolved oxygen measurement, in mg/L}
  }}
\source{
https://www80.tceq.texas.gov/SwqmisPublic/public/default.htm
}
\usage{
data(Dissolved_Oxygen)
}
\description{
Data from the Texas Commission on Environmental Quality Surface Water Quality
Monitoring Information System. The AverageDO field is the mean of dissolved oxygen
concentrations (mg/L) measured at a field site at that day. The MinDO is the minimum
dissolved oxygen concentration measured at that site on that day.
}
\keyword{datasets}
