% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Entero}
\alias{Entero}
\title{Enterococci bacteria measurements from the Tres Palacios River}
\format{
A data frame with 212 rows and 5 variables:
\describe{
  \item{Station_ID}{unique water quality monitoring station identifier}
  \item{Date}{sampling date in yyyy-mm-dd format}
  \item{Param_Code}{unique parameter code}
  \item{Param_Desc}{parameter description with units}
  \item{Value}{Enterococci concentration, in MPN/L}
  }
}
\source{
https://www80.tceq.texas.gov/SwqmisPublic/public/default.htm
}
\usage{
data(Entero)
}
\description{
Data from the Texas Commission on Environmental Quality Surface Water Quality
Monitoring Information System. The `Value`` field is the lab measured value 
of Enterococci bacteria (MPN/100 mL) from grab samples collected at 
`Station ID` on the Tres Palacios River on `Date`.
}
\keyword{datasets}
