% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_gmean.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculates the Geometric Mean}
\usage{
gm_mean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{logical TRUE/FALSE remove NA values}

\item{zero.propagate}{logical TRUE/FALSE. Allows the optional propagation of
zeros.}
}
\value{
the geometric mean of the vector
}
\description{
Originally from Paul McMurdie, Ben Bolker, and Gregor on Stack Overflow:
https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in
}
\keyword{internal}
