% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{gmean_ci}
\alias{gmean_ci}
\title{Geometric Mean and Standard Deviation}
\usage{
gmean_ci(x, method = c("classic", "boot"), conf.level = NA,
  sides = c("two.sided", "left", "right"), na.rm = FALSE, ...)
}
\arguments{
\item{x}{a positive numeric vector. An object which is not a vector is
coerced (if possible) by as.vector.}

\item{method}{a vector of character strings representing the type of
intervals required. The value should be any subset of the values
\code{"classic"}, \code{"boot"}.}

\item{conf.level}{confidence level of the interval. Default is \code{NA}.}

\item{sides}{a character string specifying the side of the confidence
interval, must be one of \code{"two.sided"} (default), \code{"left"} or
\code{"right"}. You can specify just the initial letter. \code{"left"}
would be analogue to a hypothesis of \code{"greater"} in a \code{t.test}.}

\item{na.rm}{logical, indicating whether \code{NA} values should be stripped
before the computation proceeds. Defaults to \code{FALSE}.}

\item{...}{further arguments are passed to the \code{\link{boot}} function.
Supported arguments are \code{type} (\code{"norm"}, \code{"basic"},
\code{"stud"}, \code{"perc"}, \code{"bca"}), \code{parallel} and the number
of bootstrap replicates \code{R}. If not defined those will be set to their
defaults, being \code{"basic"} for \code{type}, option
\code{"boot.parallel"} (and if that is not set, \code{"no"}) for
\code{parallel} and \code{999} for \code{R}.}
}
\value{
a numeric value.
}
\description{
An implementation of the Gmean function in Andri Signorell's DescTools
pacakage. Calculates the geometric mean, its confidence interval and the
geometric standard deviation of a vector x.
}
\seealso{
\code{\link[DescTools]{Gmean}}
}
\author{
Andri Signorell <andri@signorell.net>
}
\keyword{internal}
