% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{mean_ci}
\alias{mean_ci}
\title{Confidence Interval for the Mean}
\usage{
mean_ci(x, sd = NULL, trim = 0, method = c("classic", "boot"),
  conf.level = 0.95, sides = c("two.sided", "left", "right"),
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{sd}{the standard deviation of x. If provided it's interpreted as sd of
the population and the normal quantiles will be used for constructing the
confidence intervals. If left to \code{NULL} (default) the sample
\code{sd(x)} will be calculated and used in combination with the
t-distribution.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of \code{x} before the mean is computed. Values of \code{trim} outside
that range are taken as the nearest endpoint.}

\item{method}{A vector of character strings representing the type of
intervals required. The value should be any subset of the values
\code{"classic"}, \code{"boot"}.}

\item{conf.level}{confidence level of the interval.}

\item{sides}{a character string specifying the side of the confidence
interval, must be one of \code{"two.sided"} (default), \code{"left"} or
\code{"right"}. You can specify just the initial letter. \code{"left"}
would be analogue to a hypothesis of \code{"greater"} in a \code{t.test}.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. Defaults to FALSE.}

\item{...}{further arguments are passed to the \code{\link{boot}} function.
Supported arguments are \code{type} (\code{"norm"}, \code{"basic"},
\code{"stud"}, \code{"perc"}, \code{"bca"}), \code{parallel} and the number
of bootstrap replicates \code{R}. If not defined those will be set to their
defaults, being \code{"basic"} for \code{type},  option
\code{"boot.parallel"} (and if that is not set, \code{"no"}) for
\code{parallel}}
}
\value{
a numeric vector with 3 elements: \item{mean}{mean},
  \item{lwr.ci}{lower bound of the confidence interval}, \item{upr.ci}{upper
  bound of the confidence interval}
}
\description{
An implementation of the MeanCI function in Andri Signorell's DescTools
pacakage. Collection of several approaches to determine confidence intervals
for the mean. Both, the classical way and bootstrap intervals are implemented
for both, normal and trimmed means.
}
\author{
Andri Signorell <andri@signorell.net>
}
\keyword{internal}
