% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_window}
\alias{open_window}
\title{Open Window}
\usage{
open_window(x, tcolumn, unit = "years", n, i, na.pad)
}
\arguments{
\item{x}{dataframe}

\item{tcolumn}{time column}

\item{unit}{unit}

\item{n}{desired n}

\item{i}{row number}

\item{na.pad}{logical if `na.pad = TRUE` incomplete windows (duration of the window < `n`) return `NA`.}
}
\value{
vector
}
\description{
calculates the period at each row from the row of interest
}
\keyword{internal}
