% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_binom.R
\name{tbr_binom}
\alias{tbr_binom}
\title{Time-Based Rolling Binomial Probability}
\usage{
tbr_binom(.tbl, x, tcolumn, unit = "years", n, alpha = 0.05)
}
\arguments{
\item{.tbl}{a data frame with at least two variables; time column formatted as date, date/time and value column.}

\item{x}{column containing "success" and "failure" observations as 0 or 1}

\item{tcolumn}{formated time column}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{alpha}{numeric, probability of a type 1 error, so confidence coefficient = 1-alpha}
}
\value{
tibble with columns for the rolling binomial probability and upper and lower confidence intervals.
}
\description{
Produces a a rolling time-window based vector of binomial probability and confidence intervals.
}
\examples{
## Generate Sample Data
df <- tibble::data_frame(
date = sample(seq(as.Date('2000-01-01'), as.Date('2015/12/30'), by = "day"), 100),
value = rbinom(100, 1, 0.25)
)

## Run Function
df <- tbr_binom(df, x = value,
tcolumn = date, unit = "years", n = 5,
alpha = 0.1)
}
