% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_binom.R
\name{tbr_binom_window}
\alias{tbr_binom_window}
\title{Binomial test based on time window}
\usage{
tbr_binom_window(x, tcolumn, unit = "years", n, i, alpha)
}
\arguments{
\item{x}{column containing "success" and "failure" observations as 0 or 1}

\item{tcolumn}{formated time column}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours",
"minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{i}{rows}

\item{alpha}{numeric, probability of a type 1 error, so confidence
coefficient = 1-alpha}
}
\value{
list
}
\description{
Binomial test based on time window
}
\keyword{internal}
