% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_gmean.R
\name{tbr_gmean}
\alias{tbr_gmean}
\title{Time-Based Rolling Geometric Mean}
\usage{
tbr_gmean(.tbl, x, tcolumn, unit = "years", n, conf.level = 0.95)
}
\arguments{
\item{.tbl}{a data frame with at least two variables; time column formatted as date, date/time and value column.}

\item{x}{column containing the values to calculate the geometric mean.}

\item{tcolumn}{formated time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{conf.level}{confidence level of the interval. Default is .95.}
}
\value{
tibble with columns for the rolling geometric mean and upper and lower confidence levels.
}
\description{
Produces a a rolling time-window based vector of geometric means and confidence intervals.
}
\examples{
## Sample data
df <- tibble::data_frame(
date = sample(seq(as.Date('2000-01-01'),
as.Date('2015/12/30'), by = "day"), 100),
value = rexp(100, 0.2))

df <- tbrf::roll_gmean(df, x = value,
tcolumn = date, unit = "years", n = 5,
conf.level = 0.95)
}
