% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_gmean.R
\name{tbr_gmean_window}
\alias{tbr_gmean_window}
\title{Geometric mean based on a time-window}
\usage{
tbr_gmean_window(x, tcolumn, unit = "years", n, conf.level, i)
}
\arguments{
\item{x}{column containing the values to calculate the geometric mean.}

\item{tcolumn}{formated time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{conf.level}{confidence level of the interval. Default is .95.}

\item{i}{row}

\item{...}{additional arguments passed to DescTools::GMean()}
}
\value{
list
}
\description{
Geometric mean based on a time-window
}
\keyword{internal}
