% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_median.R
\name{tbr_median_window}
\alias{tbr_median_window}
\title{Median Based on a Time-Window}
\usage{
tbr_median_window(x, tcolumn, unit = "years", n, i, na.pad, ...)
}
\arguments{
\item{x}{column containing the values to calculate the median.}

\item{tcolumn}{formatted time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{i}{row}

\item{na.pad}{logical. If `na.pad = TRUE` incomplete windows (duration of the window < `n`) return `NA`.}

\item{...}{additional arguments passed to \code{\link{median_ci}}}
}
\value{
list
}
\description{
Median Based on a Time-Window
}
\keyword{internal}
