% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_sd.R
\name{tbr_sd_window}
\alias{tbr_sd_window}
\title{Standard Deviation Based on a Time-Window}
\usage{
tbr_sd_window(x, tcolumn, unit = "years", n, i, ...)
}
\arguments{
\item{x}{column containing the values to calculate the standard deviation.}

\item{tcolumn}{formatted time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{i}{row}

\item{...}{additional arguments passed to base::sd()}
}
\value{
numeric value
}
\description{
Standard Deviation Based on a Time-Window
}
\keyword{internal}
