% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_sum.R
\name{tbr_sum_window}
\alias{tbr_sum_window}
\title{Sum Based on a Time-Window}
\usage{
tbr_sum_window(x, tcolumn, unit = "years", n, i, na.rm)
}
\arguments{
\item{x}{column containing the values to calculate the sum.}

\item{tcolumn}{formatted time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{i}{row}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
numeric value
}
\description{
Sum Based on a Time-Window
}
\keyword{internal}
