% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbrf-package.R
\docType{package}
\name{tbrf}
\alias{tbrf-package}
\alias{tbrf}
\title{Time-Based Rolling Functions}
\description{
Provides rolling statistical functions based
on date and time windows instead of n-lagged observations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mps9506.github.io/tbrf/}
  \item Report bugs at \url{https://github.com/mps9506/tbrf/issues}
}

}
\author{
Michael Schramm
}
