% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{median_ci}
\alias{median_ci}
\title{Confidence Interval for the Median}
\usage{
median_ci(x, conf.level = 0.95, sides = c("two.sided", "left", "right"),
  na.rm = FALSE, method = c("exact", "boot"), R = 999)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{conf.level}{confidence level of the interval.}

\item{sides}{a character string specifying the side of the confidence
interval, must be one of \code{"two.sided"} (default), \code{"left"} or
\code{"right"}. You can specify just the initial letter. \code{"left"}
would be analogue to a hypothesis of \code{"greater"} in a \code{t.test}.}

\item{na.rm}{logical. Should missing values be removed? Defaults to
\code{FALSE}.}

\item{method}{defining the type of interval that should be calculated (one
out of \code{"exact"}, \code{"boot"}). Default is \code{"exact"}. See
Details.}

\item{R}{The number of bootstrap replicates. Usually this will be a single
positive integer. See \code{\link{boot.ci}} for details.}
}
\value{
a numeric vector with 3 elements: \item{median}{median},
  \item{lwr.ci}{lower bound of the confidence interval}, \item{upr.ci}{upper
  bound of the confidence interval}
}
\description{
An implementation of the \code{\link[DescTools]{MedianCI}} function in Andri
Signorell's DescTools package. Calculates the confidence interval for the
median.
}
\seealso{
\code{\link[DescTools]{MedianCI}}
}
\author{
Andri Signorell <andri@signorell.net>
}
\keyword{internal}
