% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_mean.R
\name{tbr_mean_window}
\alias{tbr_mean_window}
\title{Mean Based on a Time-Window}
\usage{
tbr_mean_window(x, tcolumn, unit = "years", n, i, ...)
}
\arguments{
\item{x}{column containing the values to calculate the mean.}

\item{tcolumn}{formated time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{i}{row}

\item{...}{additional arguments passed to MeanCI}
}
\value{
list
}
\description{
Mean Based on a Time-Window
}
\keyword{internal}
