% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_median.R
\name{tbr_median_window}
\alias{tbr_median_window}
\title{Median Based on a Time-Window}
\usage{
tbr_median_window(x, tcolumn, unit = "years", n, conf.level, i, ...)
}
\arguments{
\item{x}{column containing the values to calculate the median.}

\item{tcolumn}{formated time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}

\item{conf.level}{confidence level of the interval. Default is .95.}

\item{i}{row}

\item{...}{additional arguments passed to DescTools::MedianCI()}
}
\value{
list
}
\description{
Median Based on a Time-Window
}
\keyword{internal}
