% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbr_sum.R
\name{tbr_sum}
\alias{tbr_sum}
\title{Time-Based Rolling Sum}
\usage{
tbr_sum(.tbl, x, tcolumn, unit = "years", n)
}
\arguments{
\item{.tbl}{a data frame with at least two variables; time column formatted as date, date/time and value column.}

\item{x}{column containing the values to calculate the sum.}

\item{tcolumn}{formated time column.}

\item{unit}{character, one of "years", "months", "weeks", "days", "hours", "minutes", "seconds"}

\item{n}{numeric, describing the length of the time window.}
}
\value{
dataframe with column for the rolling sum.
}
\description{
Time-Based Rolling Sum
}
